# openapi.yaml
openapi: 3.1.0
info:
  title: Impact Database API
  version: 1.0.0
  description: API for uploading and querying hazard-related images and metadata.
servers:
  - url: http://localhost:8000
paths:
...
  /upload:
    post:
      summary: Upload an image with metadata
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - hazard_type
                - location
              properties:
                file:
                  type: string
                  format: binary
                hazard_type:
                  type: string
                location:
                  type: string
                country:
                  type: string
                manual_latitude:
                  type: number
                manual_longitude:
                  type: number
                title:
                  type: string
                title_i18n:
                  type: object
                  additionalProperties:
                    type: string
                abstract:
                  type: string
                abstract_i18n:
                  type: object
                  additionalProperties:
                    type: string
                purpose:
                  type: string
                purpose_i18n:
                  type: object
                  additionalProperties:
                    type: string
                keywords:
                  type: array
                  items:
                    type: string
                keywords_i18n:
                  type: object
                  additionalProperties:
                    type: array
                    items:
                      type: string
                metadata_language:
                  type: string
      responses:
        '200':
          description: Upload success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageMetadata'
        '400':
          description: Validation error
...
  /images/{filename}:
    get:
      summary: Get single image metadata
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: filename
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Image metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageMetadata'
        '404':
          description: Image not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ImageMetadata:
      type: object
      properties:
        filename:
          type: string
        hazard_type:
          type: string
        location:
          type: string
        country:
          type: string
        timestamp:
          type: string
          format: date-time
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        title:
          type: string
          nullable: true
        title_i18n:
          type: object
          additionalProperties:
            type: string
        abstract:
          type: string
          nullable: true
        abstract_i18n:
          type: object
          additionalProperties:
            type: string
        purpose:
          type: string
          nullable: true
        purpose_i18n:
          type: object
          additionalProperties:
            type: string
        keywords:
          type: array
          items:
            type: string
        keywords_i18n:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        metadata_language:
          type: string
      required:
        - filename
        - hazard_type
        - location
        - timestamp
        - metadata_language
...